add deleting files

+++ controllers/admin.js
@@ -2,6 +2,7 @@ const mongodb = require('mongodb');
 const mongoose = require('mongoose');
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -152,6 +153,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
 
@@ -189,7 +191,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id }) //findByIdAndDelete - mongoose
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      Product.deleteOne({ _id: prodId, userId: req.user._id }); //findByIdAndDelete - mongoose
+    })
     .then(() => {
       res.redirect('/admin/products');
     })
